Add files via upload
This is an simple Application with API, connected with Docker, where you can run it in Postman with JSON file. Write a single Comment as a  "tex1" and "text2" as a JSON file and send post request And you will get ratio Between "test1" and "text2" and you get answer how similar are they.